@@ -48,16 +48,16 @@
     "luxon": "^3.3.0",
     "mutationobserver-shim": "^0.3.7",
     "notistack": "^1.0.10",
-    "react": "^16.13.1",
+    "react": "18.3.1",
     "react-chartjs-2": "^5.0.0",
-    "react-dom": "^16.13.1",
+    "react-dom": "18.3.1",
     "react-dropzone": "^11.0.3",
     "react-ga4": "^2.1.0",
     "react-hook-form": "^7.41.5",
     "react-jwt": "^1.1.8",
     "react-leaflet": "^2.7.0",
     "react-leaflet-markercluster": "^2.0.0",
-    "react-material-ui-carousel": "^1.8.2",
+    "react-material-ui-carousel": "3.4.2",
     "react-redux": "^7.2.0",
     "react-router-dom": "^6.0.0",
     "react-slick": "^0.27.10",
@@ -102,8 +102,8 @@
     "@types/lodash": "^4.14.149",
     "@types/luxon": "^3.3.1",
     "@types/node": "^18.16.16",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@types/react": "18.3.1",
+    "@types/react-dom": "18.3.1",
     "@types/react-leaflet": "^2.5.1",
     "@types/react-redux": "^7.1.7",
     "@types/react-router-dom": "^5.1.3",
